fix(active_record): ArgumentError in AssociationRelation
This fix is specifically for `AssociationRelation`, which is the class
that you get when you chain a `where` onto an association and then try
to use methods such as `first_or_initialize` on the result.

In Rails 6, the signature of `build`, `create`, and `create!` changed.
They no longer accept `*args`. See
https://github.com/rails/rails/commit/f1c1c43022c24e27a02d698fd29e7955c10d2f03
for more details on that change.

When this gem's patch on `first_or_initialize` delegates to `create`,
for example, it tries to pass `options` as a second parameter. This
succeeded with the previous signature of these methods, but in Rails 6,
this fails as the methods no longer accept a second argument.

This commit adds another patch to `AssociationRelation` to override
these methods so that they accept the `options` argument and pass it on.