fix(associations): redefine build_record for CollectionAssociation
In Rails 6.0.4, CollectionAssociation defines its own implementation of
build_record which overrides the on inherited from the Association base
class. This breaks protected_attributes_continued's redefinition because
now CollectionAssociation is no longer using the implementation from the
super class.

I fixed this by adding a new override for build_record in
CollectionAssociation. The override is only applied for Rails >= 6.0.4 <
6.1. The implementation is just the same as the one from main.